# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.HTTPError:
    properties:
      code:
        type: integer
      error:
        type: string
    type: object
...
  models.Credentials:
    properties:
      password:
        type: string
      provider:
        type: string
      username:
        type: string
    required:
    - password
    - provider
    - username
    type: object
  models.CronJob:
    properties:
      disable:
        type: boolean
      extra_vars:
        additionalProperties: true
        type: object
      name:
        type: string
      owner:
        type: string
      schedule:
        type: string
      task:
        type: string
    type: object
...
  models.Job:
    properties:
      completed:
        type: integer
      completion_time:
        type: string
      failed:
        type: integer
      failed_reason:
        type: string
      json_data:
        additionalProperties: true
        type: object
      name:
        type: string
      owner:
        type: string
      start_time:
        type: string
      status:
        type: string
      stdout:
        type: string
    type: object
  models.JobMessage:
    properties:
      message:
        type: string
      name:
        type: string
    type: object
  models.LoginToken:
    properties:
      token:
        type: string
    type: object
  models.ResponseMessage:
    properties:
      message:
        example: runner successfully deleted
        type: string
    type: object
...
  models.Task:
    properties:
      command:
        type: string
      name:
        type: string
      runner:
        type: string
      schema:
        additionalProperties: {}
        type: object
      synchronous:
        type: boolean
    required:
    - command
    - name
    - runner
    type: object
...
info:
  contact:
    email: info@kubecode.io
    name: Kubecode Support
    url: https://www.kubecode.io/contact
  description: API Gateway for your kubernetes services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Kriten
  version: v0.3
paths:
...
  /cronjobs:
    get:
      consumes:
      - application/json
      description: List all Cronjobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CronJob'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all Cronjobs
      tags:
      - cronjobs
    post:
      consumes:
      - application/json
      description: Add a job to the cluster
      parameters:
      - description: New cronjob
        in: body
        name: cronjob
        required: true
        schema:
          $ref: '#/definitions/models.CronJob'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CronJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new job
      tags:
      - cronjobs
  /cronjobs/{name}:
    delete:
      consumes:
      - application/json
      description: Delete by CronJob ID
      parameters:
      - description: CronJob name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Delete a CronJob
      tags:
      - cronjobs
    get:
      consumes:
      - application/json
      description: Get information about a specific job
      parameters:
      - description: CronJob  name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CronJob'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get job info
      tags:
      - cronjobs
    patch:
      consumes:
      - application/json
      description: Update a cronjob in the cluster
      parameters:
      - description: Update CronJob
        in: body
        name: cronjob
        required: true
        schema:
          $ref: '#/definitions/models.CronJob'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CronJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Update a cronjob
      tags:
      - cronjobs
  /cronjobs/{name}/schema:
    get:
      consumes:
      - application/json
      description: Get task schema for the job info and input parameters
      parameters:
      - description: Task  name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get task schema
      tags:
      - jobs
...
  /jobs:
    get:
      consumes:
      - application/json
      description: List all jobs with optional filtering and pagination
      parameters:
      - description: Maximum number of jobs to return (default 100)
        in: query
        name: limit
        type: integer
      - description: Number of jobs to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by job owner
        in: query
        name: owner
        type: string
      - description: 'Filter by status: running, completed, failed'
        in: query
        name: status
        type: string
      - description: Filter by task/job name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Job'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: List all jobs
      tags:
      - jobs
  /jobs/{name}:
    get:
      consumes:
      - application/json
      description: Get information about a specific job
      parameters:
      - description: Job Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Get job info
      tags:
      - jobs
    post:
      consumes:
      - application/json
      description: Add a job to the cluster
      parameters:
      - description: Task  name
        in: path
        name: name
        required: true
        type: string
      - description: Extra vars
        in: body
        name: evars
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobMessage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Create a new job
      tags:
      - jobs
...
  /login:
    post:
      consumes:
      - application/json
      description: authenticate and generates a JWT token
      parameters:
      - description: Your credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginToken'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      summary: Authenticate users
      tags:
      - authenticate
  /refresh:
    get:
      consumes:
      - application/json
      description: Refresh time limit of a JWT token
      parameters:
      - description: JWT Token can be provided as Cookie
        in: header
        name: token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.HTTPError'
      security:
      - Bearer: []
      summary: Auth admin
      tags:
      - authenticate
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"